Created Jet.h
Jet.h was created by modifying the car.h files as they are very similar. Jet inherits from the PoweredVehicle class so I just had to change all the Car names to Jet and change the private variable to an int and change the function names very easy short fix.